original_server/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python
python-dotenv
aiosmtplib
